+++ docs/reference/workspace-layout.md
@@ -190,8 +190,10 @@ The main host-kit files are split by responsibility:
 - `CaptureOverlayWindow.swift`: AppKit `NSPanel` wrapper for capture overlay windows
 - `CaptureOverlayController.swift`: overlay window set, focus, stream preparation, and below-overlay
   capture source management
-- `FrozenFrameAuthority.swift`: ScreenCaptureKit frozen-frame stream lifecycle, latch-token
+- `FrozenFrameAuthority.swift`: ScreenCaptureKit frozen-frame stream setup, latch-token
   resolution, self-capture-safe frame gating, and fresh-frame authority decisions
+- `FrozenFrameStreamOutput.swift`: ScreenCaptureKit stream-output delegate adaptation,
+  usable-frame filtering, display-time conversion, and frame-record emission into the authority
 - `FrozenFrameContentFilterPlanner.swift`: frozen-frame display target planning, shareable-content
   cache freshness, self-capture-excluding content filter construction, and stream configuration
 - `NSScreenDisplayID.swift`: shared AppKit display-ID extraction for native capture surfaces
